+++ .env
@@ -1,0 +1,8 @@
+# database credentials
+POSTGRES_USER="postgres"
+POSTGRES_PASSWORD="pwd"
+POSTGRES_DB="users"
+
+# Node vars
+SESSION_SECRET="3f9514e1ab49476fa947d6a55492202f"
+API_KEY="5d4e608dc61b435981f512fb57729625"
