--- ProjectSourceCode/.gitignore
@@ -1,7 +1,0 @@
-src/node_modules
-src/package-lock.json
-.env
-# Ignore vim files:
-*~
-*.swp
-*.swo
