--- ProjectSourceCode/docker-compose.yaml
@@ -1,26 +1,0 @@
-version: '3.9'
-services:
-  db:
-    image: postgres:14
-    env_file: .env
-    expose:
-      - '5432'
-    volumes:
-      - group-project:/var/lib/postgresql/data
-      - ./src/init_data:/docker-entrypoint-initdb.d
-  web:
-    image: node:lts
-    user: 'node'
-    working_dir: /home/node/app
-    env_file: .env
-    environment:
-      - NODE_ENV=development
-    depends_on:
-      - db
-    ports:
-      - '3000:3000'
-    volumes:
-      - ./src:/home/node/app
-    command: 'npm start'
-volumes:
-  group-project:
