--- ProjectSourceCode/src/package.json
@@ -1,27 +1,0 @@
-{
-    "name": "<TODO: MusicMerge>",
-    "main": "src/index.js",
-    "dependencies": {
-      "express": "^4.6.1",
-      "pg-promise": "^10.11.1",
-      "body-parser": "1.20.0",
-      "express-session": "1.17.3",
-      "express-handlebars": "^7.1.2",
-      "handlebars": "^4.7.8",
-      "axios": "^1.1.3",
-      "bcrypt": "^5.1.0"
-    },
-    "devDependencies": {
-      "nodemon": "^2.0.7",
-      "mocha": "^6.2.2",
-      "chai": "^4.2.0",
-      "chai-http": "^4.3.0",
-      "npm-run-all": "^4.1.5"
-    },
-    "scripts": {
-      "prestart": "npm install",
-      "start": "nodemon index.js",
-      "test": "mocha",
-      "testandrun": "npm run prestart && npm run test && npm start"
-    }
-  }
