--- ProjectSourceCode/src/views/layouts/main.hbs
@@ -1,7 +1,0 @@
-<!-- TODO: Add the head.hbs partial here -->
-{{> head}}
-<!-- TODO: Add the nav.hbs partial here -->
-
-{{> nav}}
-{{{body}}}
-<!-- TODO: Add the footer.hbs partial here -->
