--- ProjectSourceCode/src/views/pages/discover.hbs
@@ -1,120 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-<meta charset="UTF-8">
-<meta name="viewport" content="width=device-width, initial-scale=1.0">
-<title>Event Selection</title>
-<style>
-  footer {
-  height: 50px; /* Adjust the value as needed */
-}
-  body {
-    font-family: Arial, sans-serif;
-    margin: 20px;
-    background-color: #f4f4f4;
-  }
-  h1 {
-    color: #333;
-  }
-  table {
-    width: 100%;
-    border-collapse: collapse;
-    margin-bottom: 20px;
-  }
-  td, th {
-    text-align: left;
-    padding: 8px;
-    border-bottom: 1px solid #ddd;
-  }
-  img {
-    border-radius: 5px;
-  }
-  .event-checkbox {
-    transform: scale(1.5);
-    margin-right: 10px;
-  }
-  .btn-primary {
-    background-color: #007bff;
-    color: white;
-    padding: 10px 20px;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    font-size: 16px;
-  }
-  .btn-primary:hover {
-    background-color: #0056b3;
-  }
-  .submit-btn-container {
-    text-align: center;
-    margin-top: 20px;
-  }
-</style>
-</head>
-<body>
-
-<h1>Discover Page</h1>
-<form id="eventsForm">
-<table>
-  <thead>
-      <tr>
-      <th>Name</th>
-      <th>Image</th>
-      <th>Number of Songs</th>
-      <th>Run-Time</th>
-    </tr>
-  </thead>
-  <tbody>
-    {{#each events}}
-      <tr>
-        <td>{{this.name}}</td>
-        <td>
-          {{#if this.images}}
-            {{#if this.images.[0]}}
-              <img src="{{this.images.[0].url}}" alt="Event Image" style="width:100px; height: 60px; object-fit: cover;">
-            {{else}}
-              <img src="../../views/images/default.jpg" alt="Default Image" style="width:100px; height: 60px; object-fit: cover;">
-            {{/if}}
-          {{else}}
-            <img src="../../views/images/default.jpg" alt="Default Image" style="width:100px; height: 60px; object-fit: cover;">
-          {{/if}}
-        </td>
-        <td>30</td>
-        <td>{{this.dates.start.localTime}}</td>
-        <td>
-            <input type="checkbox" class="event-checkbox" name="event" value="{{this.id}}">
-        </td>
-      </tr>
-    {{/each}}
-  </tbody>
-</table>
-<div class="submit-btn-container">
-  <button type="submit" class="btn-primary">Merge Playlists</button>
-</div>
-</form>
-
-<script>
-// JavaScript from previous example
-document.addEventListener('DOMContentLoaded', function () {
-  const MAX_CHECKED_COUNT = 2;
-  const checkboxes = document.querySelectorAll('.event-checkbox');
-
-  function updateCheckboxState() {
-    const checkedCount = document.querySelectorAll('.event-checkbox:checked').length;
-    checkboxes.forEach(checkbox => {
-      if (checkedCount >= MAX_CHECKED_COUNT) {
-        checkbox.disabled = checkbox.checked ? false : true;
-      } else {
-        checkbox.disabled = false;
-      }
-    });
-  }
-
-  checkboxes.forEach(checkbox => 
-    checkbox.addEventListener('change', updateCheckboxState)
-  );
-});
-</script>
-
-</body>
-</html>
