--- ProjectSourceCode/src/views/pages/login.hbs
@@ -1,5 +1,0 @@
-<form action="/login" method="POST">
-    <input type="text" name="username" placeholder="Username">
-    <input type="text" name="password" placeholder="Password">
-    <button type="submit">Login</button>
-</form>
