--- ProjectSourceCode/src/views/pages/logout.hbs
@@ -1,3 +1,0 @@
-<div class="container text-center">
-  {{> message }}
-</div>
