--- ProjectSourceCode/src/views/pages/register.hbs
@@ -1,12 +1,0 @@
-<form action="/register" method="POST" style="max-width: 1000px; margin: auto;">
-  <div class="mb-3">
-    <label for="exampleInputUsername" class="form-label">Username</label>
-    <input name="username" type="text" class="form-control" id="exampleInputUsername">
-    <label for="exampleInputPassword" class="form-label">Password</label>
-    <input name="password" type="password" class="form-control" id="exampleInputPassword">
-    <label for="exampleInputSpotify" class="form-label">Spotify Account Username</label>
-    <input name="spotifyusername" type="text" class="form-control" id="exampleInputSpotify">
-  </div>
-  <button type="submit" class="btn btn-primary">Register</button>
-  <p>Already have an account? <a href="/login">Login</a></p>
-</form>
