--- ProjectSourceCode/src/views/partials/footer.hbs
@@ -1,8 +1,0 @@
-<footer class="text-center text-muted w-100 mt-auto fixed-bottom">
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-kenU1KFdBIe4zVF0s0G1M5b4hcpxyD9F7jL+jjXkk+Q2h455rYXK/7HAuoJl+0I4" crossorigin="anonymous"></script>
-    <p>
-        &copy; Copyright 2024 : CSCI 3308 - Music Merge
-    </p>
-</footer>
-</body>
-</html>
