--- ProjectSourceCode/src/views/partials/message.hbs
@@ -1,5 +1,0 @@
-{{#if message}}
-  <div class="alert alert-{{#if error}}danger{{else}}success{{/if}}" role="alert">
-    {{ message }}
-  </div>
-{{/if}}
