--- ProjectSourceCode/src/views/partials/title.hbs
@@ -1,5 +1,0 @@
-{{#if first_name}}
-  <title> {{first_name}} - Music Merge </title>
-  {{else}}
-  <title> Music Merge </title>
-{{/if}}
